// src/redux/features/ClassicTemplate/types.ts
export type LanguageData = {
  language?: string;
  level?: string;
  id: string;
};
export type CertificationData = {
  year: string;
  certification: string;
  id: string;
};
export type AwardsData = {
  year: string;
  award: string;
  id: string;
};
export type PayloadDataNestedUpdate = {
  name: string;
  value: string;
  id: string;
};